fix: address PR #330 CI failures + phantom 403
Extracts GET handlers and types for /v1/agents/sessions into a new
session_queries module. The 600-line cap on sessions.rs is restored
(was 841 lines) and the items_after_statements clippy lint is gone
since MAX_SESSIONS is now declared at module scope.

The phantom 403 in get_session is resolved by switching to a two-step
lookup (matching delete_session): query by id only, then verify
tenant ownership in application code. The previous WHERE id = $1
AND tenant_id = $2 query returned 404 on cross-tenant access while
the OpenAPI annotation advertised 403.

Regenerates openapi.json and frontend/src/api/generated/schema.ts so
the openapi-sync and codegen-drift CI checks pass.

Fixes:
- clippy items_after_statements (sessions.rs:587)
- file-size cap (sessions.rs was 841 > 600)
- openapi-sync drift (GET routes missing from openapi.json)
- codegen drift (frontend schema out of sync)
- phantom 403 (get_session OpenAPI annotation mismatched behaviour)
