+++ services/control-api/src/openapi.rs
@@ -60,8 +60,8 @@ use crate::routes::{
         query::traversal::get_callees,
         agents::sessions::create_session,
         agents::sessions::delete_session,
-        agents::sessions::get_session,
-        agents::sessions::list_sessions,
+        agents::session_queries::get_session,
+        agents::session_queries::list_sessions,
         agents::events::session_events,
     ),
     components(
@@ -124,9 +124,9 @@ use crate::routes::{
             query::usages::UsagesResponse,
             agents::sessions::CreateSessionRequest,
             agents::sessions::CreateSessionResponse,
-            agents::sessions::SessionItem,
-            agents::sessions::ListSessionsResponse,
-            agents::sessions::SessionDetail,
+            agents::session_queries::SessionItem,
+            agents::session_queries::ListSessionsResponse,
+            agents::session_queries::SessionDetail,
         )
     ),
     tags(
